Rename organization and set new (DDD based) based project structure to readme
