+++ app/api-direct/pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>com.programmersonly</groupId>
+        <groupId>com.geeksacagemy</groupId>
         <artifactId>app</artifactId>
         <version>0.0.1-SNAPSHOT</version>
     </parent>
@@ -15,7 +15,7 @@
 
     <dependencies>
         <dependency>
-            <groupId>com.programmersonly</groupId>
+            <groupId>com.geeksacagemy</groupId>
             <artifactId>domain</artifactId>
         </dependency>
