+++ app/api-direct/src/main/java/com/geeksacagemy/mentorship/direct/template/RemoveTemplateRequest.java
@@ -1,0 +1,13 @@
+package com.geeksacagemy.mentorship.direct.template;
+
+import lombok.*;
+
+import java.util.UUID;
+
+@Builder
+@Getter
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+class RemoveTemplateRequest {
+  private UUID templateId;
+}
