--- app/api-direct/src/main/java/com/programmersonly/mentorship/direct/template/ConfirmTemplateRequest.java
@@ -1,13 +1,0 @@
-package com.programmersonly.mentorship.direct.template;
-
-import lombok.*;
-
-import java.util.UUID;
-
-@Builder
-@Getter
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-class ConfirmTemplateRequest {
-  private UUID templateId;
-}
