--- app/api-direct/src/main/java/com/programmersonly/mentorship/direct/template/CreateTemplateRequest.java
@@ -1,15 +1,0 @@
-package com.programmersonly.mentorship.direct.template;
-
-import lombok.*;
-
-import java.util.UUID;
-
-@Builder
-@Getter
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-class CreateTemplateRequest {
-  private UUID userId;
-  private String email;
-  private String fullName;
-}
