+++ app/application/src/main/java/com/geeksacagemy/mentorship/MentorshipApplication.java
@@ -1,0 +1,13 @@
+package com.geeksacagemy.mentorship;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MentorshipApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(MentorshipApplication.class, args);
+    }
+
+}
