+++ app/domain/src/main/java/com/geeksacagemy/mentorship/exception/ConfirmationConflictException.java
@@ -1,0 +1,8 @@
+package com.geeksacagemy.mentorship.exception;
+
+public class ConfirmationConflictException extends BusinessException {
+
+    public ConfirmationConflictException() {
+        super(ErrorCode.MS01);
+    }
+}
