+++ app/domain/src/main/java/com/geeksacagemy/mentorship/exception/ErrorCode.java
@@ -1,0 +1,16 @@
+package com.geeksacagemy.mentorship.exception;
+
+public enum ErrorCode {
+    MS01("Cannot confirm"),
+    MS02("Cannot find template");
+
+    private String value;
+
+    ErrorCode(String message) {
+        this.value = message;
+    }
+
+    public String getValue() {
+        return value;
+    }
+}
