+++ app/domain/src/main/java/com/geeksacagemy/mentorship/exception/ErrorResponse.java
@@ -1,0 +1,6 @@
+package com.geeksacagemy.mentorship.exception;
+
+public interface ErrorResponse {
+    ErrorCode getErrorCode();
+    String getMessage();
+}
