+++ app/domain/src/main/java/com/geeksacagemy/mentorship/exception/TemplateNotFoundException.java
@@ -1,0 +1,8 @@
+package com.geeksacagemy.mentorship.exception;
+
+public class TemplateNotFoundException extends BusinessException {
+
+    public TemplateNotFoundException() {
+        super(ErrorCode.MS02);
+    }
+}
