+++ app/domain/src/main/java/com/geeksacagemy/mentorship/template/TemplateServiceImpl.java
@@ -1,0 +1,46 @@
+package com.geeksacagemy.mentorship.template;
+
+import com.geeksacagemy.mentorship.exception.ConfirmationConflictException;
+import com.geeksacagemy.mentorship.exception.TemplateNotFoundException;
+import com.geeksacagemy.mentorship.template.dto.CreateTemplateDto;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import javax.transaction.Transactional;
+import java.util.UUID;
+
+@Service
+@RequiredArgsConstructor
+@Transactional
+class TemplateServiceImpl implements TemplateService {
+
+  private final TemplateRepository templateRepository;
+
+  @Override
+  public void create(CreateTemplateDto createTemplateDto) {
+    Template template = Template.builder()
+            .fullName(createTemplateDto.getFullName())
+            .email(createTemplateDto.getEmail())
+            .userId(createTemplateDto.getUserId())
+            .status(TemplateStatus.CREATED)
+            .build();
+    templateRepository.create(template);
+  }
+
+  @Override
+  public void confirm(UUID templateId) {
+    Template template = templateRepository.findById(templateId)
+            .filter(tmp -> tmp.getStatus().equals(TemplateStatus.CREATED))
+            .orElseThrow(ConfirmationConflictException::new);
+    templateRepository.confirm(template.getId());
+  }
+
+  @Override
+  public void remove(UUID templateId) {
+    // TODO Czy można usuwać wyłącznie template o statusie CREATED?????
+    Template template = templateRepository.findById(templateId)
+            .filter(tmp -> tmp.getStatus().equals(TemplateStatus.CREATED))
+            .orElseThrow(TemplateNotFoundException::new);
+    templateRepository.deleteById(template.getId());
+  }
+}
