+++ app/domain/src/main/java/com/geeksacagemy/mentorship/template/TemplateStatus.java
@@ -1,0 +1,5 @@
+package com.geeksacagemy.mentorship.template;
+
+enum TemplateStatus {
+    CREATED, CONFIRMED
+}
