--- app/domain/src/main/java/com/programmersonly/mentorship/exception/ConfirmationConflictException.java
@@ -1,8 +1,0 @@
-package com.programmersonly.mentorship.exception;
-
-public class ConfirmationConflictException extends BusinessException {
-
-    public ConfirmationConflictException() {
-        super(ErrorCode.MS01);
-    }
-}
