--- app/domain/src/main/java/com/programmersonly/mentorship/exception/ErrorResponse.java
@@ -1,6 +1,0 @@
-package com.programmersonly.mentorship.exception;
-
-public interface ErrorResponse {
-    ErrorCode getErrorCode();
-    String getMessage();
-}
