--- app/domain/src/main/java/com/programmersonly/mentorship/template/Template.java
@@ -1,20 +1,0 @@
-package com.programmersonly.mentorship.template;
-
-import lombok.*;
-
-import java.util.UUID;
-
-import static lombok.AccessLevel.PRIVATE;
-
-@Getter
-@Builder
-@AllArgsConstructor(access = PRIVATE)
-@NoArgsConstructor(access = PRIVATE)
-public class Template {
-
-    private UUID id;
-    private UUID userId;
-    private String email;
-    private String fullName;
-    private TemplateStatus status;
-}
