--- app/domain/src/main/java/com/programmersonly/mentorship/template/TemplateRepository.java
@@ -1,15 +1,0 @@
-package com.programmersonly.mentorship.template;
-
-import java.util.Optional;
-import java.util.UUID;
-
-public interface TemplateRepository {
-
-    void create(Template template);
-
-    Optional<Template> findById(UUID templateId);
-
-    void confirm(UUID templateId);
-
-    void deleteById(UUID templateId);
-}
