--- app/domain/src/main/java/com/programmersonly/mentorship/template/dto/CreateTemplateDto.java
@@ -1,16 +1,0 @@
-package com.programmersonly.mentorship.template.dto;
-
-import lombok.*;
-
-import java.util.UUID;
-
-@Getter
-@Builder
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class CreateTemplateDto {
-
-    private UUID userId;
-    private String email;
-    private String fullName;
-}
